remove sql editor theme, add session check
