+++ public/components/Main/main.tsx
@@ -22,8 +22,9 @@ import { IHttpResponse } from 'angular';
 import _ from 'lodash';
 import React from 'react';
 import { ChromeBreadcrumb, CoreStart } from '../../../../../src/core/public';
-import { ASYNC_QUERY_JOB_ENDPOINT } from '../../../common/constants';
+import { ASYNC_QUERY_JOB_ENDPOINT, POLL_INTERVAL_MS } from '../../../common/constants';
 import { AsyncQueryLoadingStatus } from '../../../common/types';
+import { getAsyncSessionId, setAsyncSessionId } from '../../../common/utils/async_query_helpers';
 import { MESSAGE_TAB_LABEL } from '../../utils/constants';
 import {
   Tree,
@@ -438,6 +439,7 @@ export class Main extends React.Component<MainProps, MainState> {
                 lang: language,
                 query: query,
                 datasource: this.state.selectedDatasource[0].label,
+                sessionId: getAsyncSessionId() ?? undefined,
               }),
             })
             .catch((error: any) => {
@@ -470,6 +472,7 @@ export class Main extends React.Component<MainProps, MainState> {
                 : '';
 
               const queryId: string = _.get(responseObj, 'queryId');
+              setAsyncSessionId(_.get(responseObj, 'sessionId', null));
 
               // clear state from previous results and start async loading
               this.setState({
@@ -494,7 +497,7 @@ export class Main extends React.Component<MainProps, MainState> {
                   clearInterval(interval);
                 }
                 this.callGetStartPolling(queries);
-              }, 2 * 1000);
+              }, POLL_INTERVAL_MS);
             }
           }
         );
