--- public/index.scss
@@ -1,142 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-
-.ace-sql-console {
-  .ace_gutter {
-    background: rgb(237, 239, 243);
-    color: rgb(0, 0, 0);
-  }
-  .ace_print-margin {
-    width: 1px;
-    background: #e8e8e8;
-  }
-  .ace_fold {
-    background-color: #6b72e6;
-  }
-  background-color: rgb(245, 247, 250);
-  color: black;
-  .ace_marker-layer {
-    .ace_active-line.ace_active-line {
-      background-color: rgb(211, 218, 230);
-    }
-    .ace_selection {
-      background: rgb(181, 213, 255);
-    }
-    .ace_step {
-      background: rgb(252, 255, 0);
-    }
-    .ace_stack {
-      background: rgb(164, 229, 101);
-    }
-    .ace_bracket {
-      margin: -1px 0 0 -1px;
-      border: 1px solid rgb(192, 192, 192);
-    }
-    .ace_active-line {
-      background: rgba(0, 0, 0, 0.07);
-    }
-    .ace_selected-word {
-      background: rgb(250, 250, 255);
-      border: 1px solid rgb(200, 200, 250);
-    }
-  }
-  .ace_cursor {
-    color: black;
-  }
-  .ace_invisible {
-    color: rgb(191, 191, 191);
-  }
-  .ace_storage {
-    color: rgb(157, 106, 242);
-  }
-  .ace_keyword {
-    color: rgb(157, 106, 242);
-  }
-  .ace_constant {
-    color: rgb(197, 6, 11);
-  }
-  .ace_constant.ace_buildin {
-    color: rgb(88, 72, 246);
-  }
-  .ace_constant.ace_language {
-    color: rgb(88, 92, 246);
-  }
-  .ace_constant.ace_library {
-    color: rgb(6, 150, 14);
-  }
-  .ace_invalid {
-    background-color: rgba(255, 0, 0, 0.1);
-    color: red;
-  }
-  .ace_support.ace_function {
-    color: rgb(60, 76, 114);
-  }
-  .ace_support.ace_constant {
-    color: rgb(6, 150, 14);
-  }
-  .ace_support.ace_type {
-    color: rgb(109, 121, 222);
-  }
-  .ace_support.ace_class {
-    color: rgb(109, 121, 222);
-  }
-  .ace_keyword.ace_operator {
-    color: rgb(104, 118, 135);
-  }
-  .ace_string {
-    color: rgb(3, 106, 7);
-  }
-  .ace_comment {
-    color: rgb(76, 136, 107);
-  }
-  .ace_comment.ace_doc {
-    color: rgb(0, 102, 255);
-  }
-  .ace_comment.ace_doc.ace_tag {
-    color: rgb(128, 159, 191);
-  }
-  .ace_constant.ace_numeric {
-    color: rgb(0, 0, 205);
-  }
-  .ace_variable {
-    color: rgb(49, 132, 149);
-  }
-  .ace_xml-pe {
-    color: rgb(104, 104, 91);
-  }
-  .ace_entity.ace_name.ace_function {
-    color: #0000a2;
-  }
-  .ace_heading {
-    color: rgb(12, 7, 255);
-  }
-  .ace_list {
-    color: rgb(185, 6, 144);
-  }
-  .ace_meta.ace_tag {
-    color: rgb(0, 22, 142);
-  }
-  .ace_string.ace_regex {
-    color: rgb(255, 0, 0);
-  }
-  .ace_gutter-active-line {
-    background-color: rgb(211, 218, 230);
-  }
-  .ace_indent-guide {
-    background: url('data:image/png;base64,iVBORw0KGgoAAAANSUhEUgAAAAEAAAACCAYAAACZgbYnAAAAE0lEQVQImWP4////f4bLly//BwAmVgd1/w11/gAAAABJRU5ErkJggg==')
-      right repeat-y;
-  }
-}
-.ace-sql-console.ace_multiselect {
-  .ace_selection.ace_start {
-    box-shadow: 0 0 3px 0px white;
-  }
-}
-.ace_editor {
-  .ace-sql-console {
-    height: 200px;
-  }
-}
