+++ public/index.ts
@@ -3,9 +3,6 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-
-import './index.scss';
-
 import { WorkbenchPlugin } from './plugin';
 
 // This exports static code and TypeScript types,
